Rename pkg/kics to pkg/runner and wire scan entrypoints to the runner package.
