+++ pkg/runner/service.go
@@ -3,7 +3,7 @@
  *
  * This product includes software developed at Datadog (https://www.datadoghq.com)  Copyright 2024 Datadog, Inc.
  */
-package kics
+package runner
 
 import (
 	"bytes"
